pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "jinja2>=3.1.6",
    "openpyxl>=3.1.5",
    "pyyaml>=6.0.3",
]
